Fixed Slap
Prevented users slapping themselves.
Fixed guildCreate not updating server count.
Applied various fixes to dashboard.

+++ commands/slap.js
@@ -37,10 +37,12 @@ class Slap extends Social {
       const cost = this.cmdDis(this.help.cost, level);
       const payMe = await this.cmdPay(message, message.author.id, cost, this.conf.botPerms);
       if (!payMe) return;  
+      const slapped = message.mentions.users.first().id;
+      const slapper = message.author.id;
+      if (slapper.id === slapped.id) throw 'Quit hitting yourself, quit hitting yourself, quit hitting yourself...';
       const msg = await message.channel.send('Finding Robin...');
-      const slapped = message.mentions.users.first().displayAvatarURL;
-      const slapper = message.author.displayAvatarURL;
-      const result = await getSlapped(slapper, slapped);
+
+      const result = await getSlapped(slapper.displayAvatarURL, slapped.displayAvatarURL);
       await message.channel.send({ files: [{ attachment: result, name: 'slapped.png' }] });
       await msg.delete();
     } catch (error) {
